Rename from StatelessComponent to FunctionComponent
