+++ src/components/code-preview/index.ts
@@ -1,0 +1,1 @@
+export * from './code-preview';
