--- src/components/codePreview/index.ts
@@ -1,1 +1,0 @@
-export * from './codePreview';
