+++ src/components/details/details.tsx
@@ -3,7 +3,7 @@ import { Description } from '../description';
 import { Title } from '../title';
 import { Types } from '../types';
 
-export const Details: React.StatelessComponent = () => (
+export const Details: React.FunctionComponent = () => (
   <div>
     <Title />
     <Types />
