+++ src/components/main/main.tsx
@@ -1,10 +1,10 @@
 import * as React from 'react';
-import { CodePreview } from '../codePreview';
+import { CodePreview } from '../code-preview';
 import { Configuration } from '../configuration';
 import { Details } from '../details';
 import * as styles from './main.sytes';
 
-export const Main: React.StatelessComponent = () => (
+export const Main: React.FunctionComponent = () => (
   <div className={styles.main}>
     <Details />
     <CodePreview />
