+++ src/components/title/title.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import * as styles from './title.styles';
 
-export const Title: React.StatelessComponent = () => (
+export const Title: React.FunctionComponent = () => (
   <h1 className={styles.title}>max-line-length</h1>
 );
